moves videoModule to library
