+++ libraries/video/constants/vendorCodes.js
