+++ libraries/video/shared/vastXmlBuilder.js
